requirements.txt
albumentations==1.3.1
efficientnet-pytorch==0.7.1
huggingface-hub==0.17.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
numpy==1.24.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
pandas==2.1.1
Pillow==9.3.0
pretrainedmodels==0.7.4
rasterio==1.3.8
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.3
segmentation-models-pytorch==0.3.3
timm==0.9.2
tokenizers==0.13.3
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.2.0
torchvision==0.15.2+cu118
tqdm==4.66.1
transformers==4.33.3
triton==2.0.0
